apps/bertopic_trainer/requirements.txt
jupyter
pandas
scikit-learn
plotly
matplotlib
pymongo
pydantic
torch
octis==1.12.1
bertopic
sentence-transformers
gensim
PyYAML==5.1.2